Move common benchmark code to Icicle.Benchmark
