+++ icicle.cabal
@@ -13,31 +13,35 @@ description:           icicle.
 library
   build-depends:
                        base                            >= 3          && < 5
-                     , aeson                           == 0.8.*
                      , ambiata-jetski
                      , ambiata-p
                      , ambiata-x-either
                      , ambiata-x-exception
+                     , aeson                           == 0.8.*
+                     , annotated-wl-pprint             == 0.7.*
                      , bifunctors                      == 4.2.*
                      , bytestring                      == 0.10.*
                      , containers                      == 0.5.*
                      , dates                           == 0.2.*
-                     , exceptions                      == 0.8.*
-                     , time                            == 1.4.*
-                     , lens                            == 4.7.*
-                     , old-locale                      == 1.0.*
+                     , directory                       == 1.2.*
                      , either                          == 4.3.*
+                     , exceptions                      == 0.8.*
+                     , file-embed                      == 0.0.9
                      , filepath                        == 1.3.*
+                     , lens                            == 4.7.*
+                     , mmorph                          == 1.0.*
+                     , old-locale                      == 1.0.*
+                     , parallel                        == 3.2.*
                      , scientific                      == 0.3.*
                      , semigroups                      == 0.16.*
-                     , mmorph                          == 1.0.*
+                     , temporary                       == 1.2.*
                      , text                            == 1.2.*
+                     , time                            == 1.4.*
                      , transformers                    >= 0.3        && < 0.5
+                     , unix                            == 2.7.*
                      , unordered-containers            == 0.2.*
                      , validation                      == 0.5.*
                      , vector                          == 0.10.*
-                     , annotated-wl-pprint             == 0.7.*
-                     , file-embed                      == 0.0.9
 
                      -- Well, it might look very odd that we're using both attoparsec
                      -- and parsec.
@@ -48,8 +52,6 @@ library
                      -- As an added benefit, parsec will give us nicer error messages.
                      , attoparsec                      == 0.12.*
                      ,     parsec                      == 3.1.*
-
-                     , unix                            == 2.7.*
 
   ghc-options:
                        -Wall
@@ -80,6 +82,7 @@ library
                        Icicle.Avalanche.Prim.Eval
                        Icicle.Avalanche.ToJava
 
+                       Icicle.Benchmark
                        Icicle.BubbleGum
 
                        Icicle.Common.Annot
